chore: move dependencies from handler function
